Fix issue where an iterator could be left unfreed by a failure in its configuration
The iterator will be created, but because the initialiser of GitPointer doesn't know that the failure came from a secondary setup process, doesn't know that the iterator got created in the first instance. This is fixed by splitting this up into two steps, the first to create the opaque pointer and a second to configure it. If the configuration fails, then the provided free function will be called to free the memory of the successful creation.